requirements.txt
discord.py==2.2.3
python-dotenv==0.20.0
PyYAML==6.0
GoogleBard==1.2.1